Migrate Snackbar to styled-component
