+++ src/snackbar/index.tsx
@@ -1,72 +1,40 @@
-import React, { PureComponent } from 'react'
-import { canUseDOM } from 'exenv'
-import cc from 'classcat'
-import { CrossIcon } from 'icon'
-import { createPortal } from 'react-dom'
-import { color } from '_utils/branding'
-import TransitionGroup from 'react-transition-group/TransitionGroup'
-import CustomTransition, { AnimationType } from 'transitions'
-import Button from 'button'
-import Text, { TextDisplayType } from 'text'
+import styled from 'styled-components'
+import { color, space } from '_utils/branding'
+import Snackbar from './Snackbar'
 
-import style from './style'
-
-export interface SnackbarProps {
-  readonly close: () => void
-  readonly isOpen: boolean
-  readonly className?: string
-}
-
-class Snackbar extends PureComponent<SnackbarProps> {
-  private portalNode: HTMLElement
-  static defaultProps: Partial<SnackbarProps> = {
-    isOpen: false,
+const StyledSnackbar = styled(Snackbar)`
+  & {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
   }
 
-  constructor(props: SnackbarProps) {
-    super(props)
-    if (canUseDOM) {
-      this.portalNode = document.createElement('div')
-      document.body.appendChild(this.portalNode)
-    }
+  & .kirk-snackbar {
+    position: fixed;
+    display: flex;
+    flex-direction: row;
+    width: 100%;
+    max-width: 614px;
+    margin: 0 auto;
+    bottom: 0;
+    overflow: hidden;
+    background-color: ${color.danger};
+    padding: ${space.l} ${space.xl};
+    z-index: 1000; /* above the modal */
   }
 
-  componentWillUnmount() {
-    if (this.portalNode) {
-      document.body.removeChild(this.portalNode)
-      this.portalNode = null
-    }
+  & .kirk-snackbar-content {
+    flex: 1;
+    margin: 0;
+    color: ${color.white};
   }
 
-  render() {
-    const modalElement = (
-      <TransitionGroup component="div" className="transition-wrapper kirk-snackbar-container">
-        {this.props.isOpen && (
-          <CustomTransition animationName={AnimationType.SLIDE_UP}>
-            <div className={cc(['kirk-snackbar', this.props.className])}>
-              <Text role="alert" display={TextDisplayType.TITLE} className="kirk-snackbar-content">
-                {this.props.children}
-              </Text>
-              <Button
-                status={Button.STATUS.UNSTYLED}
-                className="kirk-snackbar-cross"
-                onClick={this.props.close}
-              >
-                <CrossIcon iconColor={color.white} size="18px" />
-              </Button>
-              <style jsx>{style}</style>
-            </div>
-          </CustomTransition>
-        )}
-      </TransitionGroup>
-    )
-
-    if (!this.portalNode) {
-      return modalElement
-    }
-
-    return createPortal(modalElement, this.portalNode)
+  & .kirk-snackbar-cross {
+    padding: 0 0 0 ${space.m};
+    display: flex;
+    flex-direction: column;
+    algin-items: flex-start;
   }
-}
+`
 
-export default Snackbar
+export default StyledSnackbar
